staking: break out component impl
