+++ crates/core/component/stake/src/component/mod.rs
@@ -1,4 +1,3 @@
-// Implementation of a pd component for the staking system.
 use penumbra_distributions::component::StateReadExt as _;
 use penumbra_sct::{
     component::clock::{EpochManager, EpochRead},
@@ -9,7 +8,6 @@ use std::{
     future::Future,
     pin::Pin,
     str::FromStr,
-    sync::Arc,
 };
 use validator::BondingState::*;
 
@@ -18,9 +16,12 @@ pub mod metrics;
 pub mod rpc;
 pub use self::metrics::register_metrics;
 
+mod stake;
+
+pub use stake::Staking;
+
 use anyhow::{anyhow, Context, Result};
 use async_trait::async_trait;
-use cnidarium_component::Component;
 use futures::{FutureExt, StreamExt, TryStreamExt};
 use penumbra_asset::{asset, STAKING_TOKEN_ASSET_ID};
 
@@ -31,10 +32,7 @@ use penumbra_shielded_pool::component::{SupplyRead, SupplyWrite};
 use sha2::{Digest, Sha256};
 use tendermint::validator::Update;
 use tendermint::{
-    abci::{
-        self,
-        types::{CommitInfo, Misbehavior},
-    },
+    abci::types::{CommitInfo, Misbehavior},
     block, PublicKey,
 };
 use tokio::task::JoinSet;
@@ -69,9 +67,6 @@ fn validator_address(ck: &PublicKey) -> [u8; 20] {
 
     addr
 }
-
-// Staking component
-pub struct Staking {}
 
 pub trait ValidatorUpdates: StateRead {
     /// Returns a list of validator updates to send to Tendermint.
@@ -1116,141 +1111,6 @@ pub(crate) trait StakingImpl: StateWriteExt {
 }
 
 impl<T: StateWrite + StateWriteExt + ?Sized> StakingImpl for T {}
-
-#[async_trait]
-impl Component for Staking {
-    type AppState = (
-        crate::genesis::Content,
-        penumbra_shielded_pool::genesis::Content,
-    );
-
-    #[instrument(name = "staking", skip(state, app_state))]
-    async fn init_chain<S: StateWrite>(mut state: S, app_state: Option<&Self::AppState>) {
-        match app_state {
-            Some((staking_genesis, sp_genesis)) => {
-                state.put_stake_params(staking_genesis.stake_params.clone());
-
-                let starting_height = state
-                    .get_block_height()
-                    .await
-                    .expect("should be able to get initial block height");
-                let starting_epoch = state
-                    .get_epoch_by_height(starting_height)
-                    .await
-                    .expect("should be able to get initial epoch");
-                let epoch_index = starting_epoch.index;
-
-                let genesis_base_rate = BaseRateData {
-                    epoch_index,
-                    base_reward_rate: 0u128.into(),
-                    base_exchange_rate: 1_0000_0000u128.into(),
-                };
-                state.set_base_rate(genesis_base_rate.clone());
-
-                // Compile totals of genesis allocations by denom, which we can use
-                // to compute the delegation tokens for each validator.
-                let mut genesis_allocations = BTreeMap::<_, Amount>::new();
-                for allocation in &sp_genesis.allocations {
-                    let value = allocation.value();
-                    *genesis_allocations.entry(value.asset_id).or_default() += value.amount;
-                }
-
-                // Add initial validators to the JMT
-                // Validators are indexed in the JMT by their public key,
-                // and there is a separate key containing the list of all validator keys.
-                for validator in &staking_genesis.validators {
-                    // Parse the proto into a domain type.
-                    let validator = Validator::try_from(validator.clone())
-                        .expect("should be able to parse genesis validator");
-
-                    state
-                        .add_genesis_validator(&genesis_allocations, &genesis_base_rate, validator)
-                        .await
-                        .expect("should be able to add genesis validator to state");
-                }
-
-                // First, "prime" the state with an empty set, so the build_ function can read it.
-                state.put(
-                    state_key::current_consensus_keys().to_owned(),
-                    CurrentConsensusKeys::default(),
-                );
-
-                // Finally, record that there were no delegations in this block, so the data
-                // isn't missing when we process the first epoch transition.
-                state
-                    .set_delegation_changes(
-                        starting_height
-                            .try_into()
-                            .expect("should be able to convert u64 into block height"),
-                        Default::default(),
-                    )
-                    .await;
-            }
-            None => { /* perform upgrade specific check */ }
-        }
-        // Build the initial validator set update.
-        state
-            .build_tendermint_validator_updates()
-            .await
-            .expect("should be able to build initial tendermint validator updates");
-    }
-
-    #[instrument(name = "staking", skip(state, begin_block))]
-    async fn begin_block<S: StateWrite + 'static>(
-        state: &mut Arc<S>,
-        begin_block: &abci::request::BeginBlock,
-    ) {
-        let state = Arc::get_mut(state).expect("state should be unique");
-        // For each validator identified as byzantine by tendermint, update its
-        // state to be slashed. If the validator is not tracked in the JMT, this
-        // will be a no-op. See #2919 for more details.
-        for evidence in begin_block.byzantine_validators.iter() {
-            let _ = state.process_evidence(evidence).await.map_err(|e| {
-                tracing::warn!(?e, "failed to process byzantine misbehavior evidence")
-            });
-        }
-
-        state
-            .track_uptime(&begin_block.last_commit_info)
-            .await
-            .expect("should be able to track uptime");
-    }
-
-    #[instrument(name = "staking", skip(state, end_block))]
-    async fn end_block<S: StateWrite + 'static>(
-        state: &mut Arc<S>,
-        end_block: &abci::request::EndBlock,
-    ) {
-        let state = Arc::get_mut(state).expect("state should be unique");
-        // Write the delegation changes for this block.
-        state
-            .set_delegation_changes(
-                end_block
-                    .height
-                    .try_into()
-                    .expect("should be able to convert i64 into block height"),
-                state.get_delegation_changes().clone(),
-            )
-            .await;
-    }
-
-    #[instrument(name = "staking", skip(state))]
-    async fn end_epoch<S: StateWrite + 'static>(state: &mut Arc<S>) -> anyhow::Result<()> {
-        let state = Arc::get_mut(state).context("state should be unique")?;
-        let epoch_ending = state
-            .get_current_epoch()
-            .await
-            .context("should be able to get current epoch during end_epoch")?;
-        state.end_epoch(epoch_ending).await?;
-        // Since we only update the validator set at epoch boundaries,
-        // we only need to build the validator set updates here in end_epoch.
-        state
-            .build_tendermint_validator_updates()
-            .await
-            .context("should be able to build tendermint validator updates")?;
-        Ok(())
-    }
-}
 
 /// Extension trait providing read access to staking data.
 #[async_trait]
